requirements.txt
beautifulsoup4==4.9.3
faster_whisper==1.1.1
Flask==3.1.1
gTTS==2.5.4
keras==3.11.0
langchain==0.3.27
langchain_community==0.3.27
langdetect==1.0.9
language_tool_python==2.9.4
numpy==2.2.6
playsound==1.2.2
playwright==1.53.0
pyspellchecker==0.8.3
python-dotenv==1.1.1
qdrant_client==1.15.1
Requests==2.32.4
scipy==1.16.1
selenium==4.34.2
sentence_transformers==5.0.0
sounddevice==0.5.2
tensorflow==2.19.0
together==1.5.23
transformers==4.46.2
webdriver_manager==4.0.2
tf-keras==2.19.0
opencv-python==4.12.0.88
librosa==0.11.0
